refact - seperate client&remote endpoint to tcp&udp
